# Optimize use of decryption request WorkerPool