--- a/newsfragments/3393.dev.rst
+++ b/newsfragments/3393.dev.rst
@@ -1,0 +1,1 @@
+Optimize use of decryption request WorkerPool.